All test files should use a new independant instance of convict
